- Updated dependencies - Refactored webpack-permissions plugin
