+++ package.json
@@ -52,50 +52,49 @@
   },
   "homepage": "https://github.com/first-line-outsourcing/flo.sls#readme",
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.61",
+    "@types/aws-lambda": "^8.10.63",
     "@types/bluebird": "^3.5.32",
-    "@types/jest": "^25.2.1",
-    "@types/node": "^13.13.15",
+    "@types/jest": "^26.0.14",
+    "@types/node": "^14.11.8",
     "@types/validator": "^13.1.0",
-    "@typescript-eslint/eslint-plugin": "^3.1.0",
-    "@typescript-eslint/parser": "^3.1.0",
-    "copy-webpack-plugin": "^6.1.0",
-    "dotenv-webpack": "^1.7.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-prettier": "^3.1.3",
-    "husky": "^4.2.5",
-    "jest": "^25.5.4",
-    "prettier": "^2.0.5",
-    "pretty-quick": "^2.0.1",
-    "serverless": "^1.70.1",
-    "serverless-dir-config-plugin": "0.1.5",
-    "serverless-env-generator": "1.4.1",
-    "serverless-offline": "5.12.1",
-    "serverless-offline-sns": "0.68.0",
-    "serverless-offline-sqs": "3.1.3",
-    "serverless-prune-plugin": "1.4.2",
-    "serverless-s3-local": "0.6.1",
-    "serverless-webpack": "5.3.2",
-    "sonarqube-scanner": "^2.6.0",
+    "@typescript-eslint/eslint-plugin": "^4.4.0",
+    "@typescript-eslint/parser": "^4.4.0",
+    "copy-webpack-plugin": "^6.2.0",
+    "dotenv-webpack": "~2.0.0",
+    "eslint": "^7.10.0",
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.3.0",
+    "jest": "^26.5.2",
+    "prettier": "^2.1.2",
+    "pretty-quick": "^3.0.2",
+    "serverless": "^2.6.0",
+    "serverless-dir-config-plugin": "^0.1.5",
+    "serverless-env-generator": "^1.4.1",
+    "serverless-offline": "^6.8.0",
+    "serverless-offline-sns": "^0.68.0",
+    "serverless-offline-sqs": "^4.1.1",
+    "serverless-prune-plugin": "^1.4.3",
+    "serverless-s3-local": "^0.6.7",
+    "serverless-webpack": "^5.3.5",
+    "sonarqube-scanner": "^2.7.0",
     "sonarqube-verify": "^1.0.2",
-    "ts-jest": "^25.5.0",
-    "ts-loader": "^7.0.2",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typescript": "^3.8.3",
+    "ts-jest": "^26.4.1",
+    "ts-loader": "^8.0.4",
+    "tsconfig-paths-webpack-plugin": "^3.3.0",
+    "typescript": "^4.0.3",
     "webpack": "^4.43.0",
-    "webpack-node-externals": "^1.7.2",
-    "webpack-shell-plugin-next": "^1.1.9"
+    "webpack-node-externals": "^2.5.2"
   },
   "dependencies": {
     "alai": "^1.0.3",
-    "aws-sdk": "^2.745.0",
+    "aws-sdk": "^2.770.0",
     "axios": "^0.20.0",
     "dotenv": "^8.2.0",
     "dynamoose": "^2.3.0",
     "node-uuid": "^1.4.8",
-    "nodemailer": "^6.4.11",
+    "nodemailer": "^6.4.13",
     "reflect-metadata": "^0.1.13",
     "sequelize": "^5.22.3",
     "sequelize-typescript": "^1.1.0",
